requirements.txt
certifi==2024.8.30
fastapi==0.115.6
pandas==2.2.3
passlib==1.7.4
pydantic==2.10.4
pymongo==4.7.1
python-dotenv==1.0.1
python_jose==3.3.0
uvicorn
python-multipart